requirements.txt
asgiref==3.7.2
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.1
djangorestframework==3.14.0
executing==1.2.0
iniconfig==2.0.0
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
pytest==7.3.1
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
